feat: Replace V1 UI with V2 and update all tests
