+++ service/static/js/app.js
@@ -1,376 +1,902 @@
-$(function () {
+// service/static/v2/js/app-v2.js
+(function () {
+  'use strict';
 
-    // ****************************************
-    //  U T I L I T Y   F U N C T I O N S
-    // ****************************************
+  /* -------------------------
+     Helpers
+  ------------------------- */
+  function $id(id) { return document.getElementById(id); }
 
-    const $flash = $("#flash_message");
-    const $searchResults = $("#search_results");
+  function escapeHtml(str) {
+    if (str === null || str === undefined) return '';
+    return String(str)
+      .replaceAll('&', '&amp;').replaceAll('<', '&lt;').replaceAll('>', '&gt;')
+      .replaceAll('"', '&quot;').replaceAll("'", '&#39;');
+  }
 
-    function handleFail(res, fallback) {
-        const message = res.responseJSON?.message || res.responseText || fallback || "Server error";
-        flashMessage(message);
+  function formatDateShort(d) {
+    if (!d) return '';
+    try { return String(d).slice(0, 10); } catch (e) { return String(d); }
+  }
+
+  function normalizeResponse(json) {
+    if (Array.isArray(json)) return json;
+    if (!json) return [];
+    if (Array.isArray(json.promotions)) return json.promotions;
+    if (Array.isArray(json.data)) return json.data;
+    if (Array.isArray(json.payload)) return json.payload;
+    for (var k in json) {
+      if (Array.isArray(json[k])) return json[k];
+    }
+    return [];
+  }
+
+  /* -------------------------
+     Table rendering / loader
+  ------------------------- */
+  // Global filter state
+  var currentFilters = {
+    active: null,      // null, 'all', 'active', 'inactive'
+    name: null,
+    type: null,
+    productId: null
+  };
+
+  async function loadPromotions(queryParams) {
+    var tbody = document.querySelector('#promotions_table tbody');
+    if (!tbody) return;
+    tbody.innerHTML = '<tr><td colspan="8" class="text-center small text-muted">Loading...</td></tr>';
+
+    try {
+      var url = '/promotions';
+      if (queryParams) {
+        url += '?' + queryParams;
+      }
+
+      var res = await fetch(url, { credentials: 'same-origin' });
+      if (!res.ok) throw new Error('GET /promotions failed: ' + res.status);
+      var json = await res.json();
+      var items = normalizeResponse(json);
+      renderRows(items);
+      updatePromotionsCount(items.length);
+    } catch (err) {
+      console.error(err);
+      tbody.innerHTML = '<tr><td colspan="8" class="text-danger">Failed to load promotions. See console.</td></tr>';
+      updatePromotionsCount(0);
+    }
+  }
+
+  function updatePromotionsCount(count) {
+    var countEl = $id('promotionsCount');
+    if (countEl) {
+      countEl.textContent = count + ' promotion' + (count !== 1 ? 's' : '');
+    }
+  }
+
+  function renderRows(items) {
+    var tbody = document.querySelector('#promotions_table tbody');
+    if (!tbody) return;
+    tbody.innerHTML = '';
+    if (!items || items.length === 0) {
+      tbody.innerHTML = '<tr><td colspan="8" class="text-center text-muted">No promotions available</td></tr>';
+      var info0 = $id('page_info');
+      if (info0) info0.textContent = '0 promotions';
+      return;
+    }
+    var frag = document.createDocumentFragment();
+    items.forEach(function (p) {
+      var tr = document.createElement('tr');
+      var id = escapeHtml(p.id ?? p.promotion_id ?? '');
+      var name = escapeHtml(p.name ?? p.title ?? '');
+      var type = escapeHtml(p.promotion_type ?? p.type ?? '');
+      var value = escapeHtml((p.value === 0 || p.value) ? p.value : (p.discount ?? ''));
+      var productId = escapeHtml(p.product_id ?? p.sku ?? '');
+      var startDate = escapeHtml(formatDateShort(p.start_date ?? p.start ?? p.starts_at));
+      var endDate = escapeHtml(formatDateShort(p.end_date ?? p.end ?? p.ends_at));
+
+      tr.innerHTML = [
+        '<td class="text-center">' + id + '</td>',
+        '<td>' + name + '</td>',
+        '<td>' + type + '</td>',
+        '<td class="text-center">' + value + '</td>',
+        '<td class="text-center">' + productId + '</td>',
+        '<td class="text-center">' + startDate + '</td>',
+        '<td class="text-center">' + endDate + '</td>',
+        '<td class="text-center" style="white-space: nowrap;">' +
+        '<button class="edit-btn" data-id="' + id + '" data-promotion=\'' + JSON.stringify(p) + '\'>' +
+        '<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-pencil" viewBox="0 0 16 16">' +
+        '<path d="M12.146.146a.5.5 0 0 1 .708 0l3 3a.5.5 0 0 1 0 .708l-10 10a.5.5 0 0 1-.168.11l-5 2a.5.5 0 0 1-.65-.65l2-5a.5.5 0 0 1 .11-.168l10-10zM11.207 2.5 13.5 4.793 14.793 3.5 12.5 1.207 11.207 2.5zm1.586 3L10.5 3.207 4 9.707V10h.5a.5.5 0 0 1 .5.5v.5h.5a.5.5 0 0 1 .5.5v.5h.293l6.5-6.5zm-9.761 5.175-.106.106-1.528 3.821 3.821-1.528.106-.106A.5.5 0 0 1 5 12.5V12h-.5a.5.5 0 0 1-.5-.5V11h-.5a.5.5 0 0 1-.468-.325z"/>' +
+        '</svg>' +
+        '</button><button class="delete-btn" data-id="' + id + '" data-name="' + name + '">' +
+        '<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-trash" viewBox="0 0 16 16">' +
+        '<path d="M5.5 5.5A.5.5 0 0 1 6 6v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5m2.5 0a.5.5 0 0 1 .5.5v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5m3 .5a.5.5 0 0 0-1 0v6a.5.5 0 0 0 1 0z"/>' +
+        '<path d="M14.5 3a1 1 0 0 1-1 1H13v9a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V4h-.5a1 1 0 0 1-1-1V2a1 1 0 0 1 1-1H6a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1h3.5a1 1 0 0 1 1 1zM4.118 4 4 4.059V13a1 1 0 0 0 1 1h6a1 1 0 0 0 1-1V4.059L11.882 4zM2.5 3h11V2h-11z"/>' +
+        '</svg>' +
+        '</button>' +
+        '</td>'
+      ].join('');
+      frag.appendChild(tr);
+    });
+    tbody.appendChild(frag);
+
+    var info = $id('page_info');
+    if (info) info.textContent = items.length + ' promotions';
+  }
+
+  /* -------------------------
+     Create modal: helpers
+  ------------------------- */
+  function normalizeDateInput(s) {
+    if (!s) return null;
+    var t = String(s).trim().replace(/\//g, '-');
+    if (/^\d{4}-\d{2}-\d{2}$/.test(t)) return t;
+    // try Date parse
+    var d = new Date(t);
+    if (!isNaN(d.getTime())) {
+      var y = d.getFullYear();
+      var m = String(d.getMonth() + 1).padStart(2, '0');
+      var day = String(d.getDate()).padStart(2, '0');
+      return y + '-' + m + '-' + day;
+    }
+    return null;
+  }
+
+  function showCreateFieldError(fieldId, message) {
+    var el = $id(fieldId);
+    if (!el) return;
+    el.classList.add('is-invalid');
+    // find the next .invalid-feedback
+    var fb = el.nextElementSibling;
+    if (fb && fb.classList && fb.classList.contains('invalid-feedback')) {
+      fb.textContent = message;
+    } else {
+      // fallback: global error box
+      var ce = $id('createError');
+      if (ce) { ce.classList.remove('d-none'); ce.textContent = message; }
+    }
+  }
+
+  function clearCreateFieldErrors() {
+    var ids = ['inputProductId', 'inputValue', 'inputName', 'inputType', 'inputStart', 'inputEnd'];
+    ids.forEach(function (id) {
+      var el = $id(id);
+      if (el) {
+        el.classList.remove('is-invalid');
+      }
+    });
+    var ce = $id('createError');
+    if (ce) { ce.classList.add('d-none'); ce.textContent = ''; }
+  }
+
+  function showSuccessToast(msg) {
+    try {
+      var toast = document.createElement('div');
+      toast.className = 'alert alert-success position-fixed';
+      // position top-right
+      toast.style.top = '1rem';
+      toast.style.right = '1rem';
+      toast.style.zIndex = 12000;
+      toast.textContent = msg || 'Success';
+      document.body.appendChild(toast);
+      setTimeout(function () { toast.remove(); }, 3000);
+    } catch (e) {
+      console.log('toast failed', e);
+    }
+  }
+
+  /* -------------------------
+     Initialize: load table
+  ------------------------- */
+  document.addEventListener('DOMContentLoaded', function () {
+    loadPromotions();
+  });
+
+  /* -------------------------
+     Create modal: logic
+     (kept in same IIFE so loadPromotions is visible)
+  ------------------------- */
+  (function initCreateModal() {
+    var createModalEl = $id('createModal');
+    var createModal = null;
+    if (createModalEl && window.bootstrap && window.bootstrap.Modal) {
+      try {
+        createModal = new bootstrap.Modal(createModalEl, {});
+      } catch (e) {
+        console.warn('Bootstrap Modal init failed', e);
+      }
     }
 
-    function flashMessage(message) {
-        $flash.empty();
-        if (message) {
-            $flash.append(message);
+    // auto-focus first input when modal shown
+    if (createModalEl) {
+      createModalEl.addEventListener('shown.bs.modal', function () {
+        var el = $id('inputName');
+        if (el) el.focus();
+      });
+      createModalEl.addEventListener('hidden.bs.modal', function () {
+        // clear the form when modal is hidden
+        var form = $id('createForm');
+        if (form) {
+          form.reset();
+          form.classList.remove('was-validated');
         }
+        clearCreateFieldErrors();
+      });
     }
 
-    // Updates the form with data from the response
-    function update_form_data(res) {
-        $("#promotion_id").val(res.id ?? "");
-        $("#promotion_name").val(res.name ?? "");
-        $("#promotion_promotion_type").val(res.promotion_type ?? "");
-        $("#promotion_value").val(res.value ?? "");
-        $("#promotion_product_id").val(res.product_id ?? "");
-        $("#promotion_start_date").val(res.start_date ?? "");
-        $("#promotion_end_date").val(res.end_date ?? "");
-    }
+    var createForm = $id('createForm');
+    if (!createForm) return;
 
-    /// Clears all form fields
-    function clear_form_data() {
-        $("#promotion_id").val("");
-        $("#promotion_name").val("");
-        $("#promotion_promotion_type").val("");
-        $("#promotion_value").val("");
-        $("#promotion_product_id").val("");
-        $("#promotion_start_date").val("");
-        $("#promotion_end_date").val("");
-    }
+    createForm.addEventListener('submit', async function (ev) {
+      ev.preventDefault();
+      ev.stopPropagation();
 
-    function renderPromotionTable(promotions) {
-        $searchResults.empty();
-        if (!promotions || promotions.length === 0) {
-            $searchResults.append("<p>No promotions found.</p>");
-            return;
+      createForm.classList.add('was-validated');
+
+      // basic HTML5 validity
+      if (!createForm.checkValidity()) {
+        return;
+      }
+
+      // gather and coerce fields
+      clearCreateFieldErrors();
+
+      var rawName = ($id('inputName').value || '').trim();
+      var rawType = ($id('inputType').value || '').trim();
+      var rawValue = $id('inputValue').value;
+      var rawProductId = ($id('inputProductId').value || '').trim();
+      var rawStart = $id('inputStart').value || '';
+      var rawEnd = $id('inputEnd').value || '';
+
+      // product_id: if provided, must be integer
+      var productId = null;
+      if (rawProductId !== '') {
+        var parsedPid = parseInt(rawProductId, 10);
+        if (Number.isNaN(parsedPid) || !Number.isFinite(parsedPid)) {
+          showCreateFieldError('inputProductId', 'Product ID must be an integer');
+          return;
         }
+        productId = parsedPid;
+      } else {
+        productId = null;
+      }
 
-        let table = '<table class="table table-striped" cellpadding="10">';
-        table += "<thead><tr>";
-        table += '<th class="col-md-2">ID</th>';
-        table += '<th class="col-md-2">Name</th>';
-        table += '<th class="col-md-2">Type</th>';
-        table += '<th class="col-md-2">Value</th>';
-        table += '<th class="col-md-2">Product ID</th>';
-        table += '<th class="col-md-2">Start Date</th>';
-        table += '<th class="col-md-2">End Date</th>';
-        table += "</tr></thead><tbody>";
+      // value coercion
+      var valueNum = Number(rawValue);
+      if (!Number.isFinite(valueNum) || valueNum < 0) {
+        showCreateFieldError('inputValue', 'Value must be a non-negative number');
+        return;
+      }
 
-        promotions.forEach((promo, index) => {
-            table += `<tr id="row_${index}"><td>${promo.id ?? ""}</td><td>${promo.name ?? ""}</td><td>${promo.promotion_type ?? ""}</td><td>${promo.value ?? ""}</td><td>${promo.product_id ?? ""}</td><td>${promo.start_date ?? ""}</td><td>${promo.end_date ?? ""}</td></tr>`;
+      // normalize dates
+      var startDate = normalizeDateInput(rawStart);
+      var endDate = normalizeDateInput(rawEnd);
+      // If user typed an unparsable date, leave null (backend will validate)
+      // or if you want to enforce: showCreateFieldError('inputStart', 'Invalid date');
+      // but we'll allow null and rely on server checks for requiredness.
+
+      var payload = {
+        name: rawName || null,
+        promotion_type: rawType || null,
+        value: valueNum,
+        product_id: productId,
+        start_date: startDate,
+        end_date: endDate
+      };
+
+      // UI: disable submit
+      var submitBtn = $id('createSubmit');
+      if (submitBtn) {
+        submitBtn.disabled = true;
+        submitBtn.textContent = 'Creating...';
+      }
+      // mark form busy
+      createForm.setAttribute('aria-busy', 'true');
+
+      try {
+        var res = await fetch('/promotions', {
+          method: 'POST',
+          headers: { 'Content-Type': 'application/json' },
+          credentials: 'same-origin',
+          body: JSON.stringify(payload)
         });
 
-        table += "</tbody></table>";
-        $searchResults.append(table);
-        if (promotions.length > 0) {
-            update_form_data(promotions[0]);
+        if (!res.ok) {
+          // try to parse structured errors
+          var errText = '';
+          try {
+            var errJson = await res.json();
+            // if backend returns {errors: {product_id: 'must be integer'}} or {message: '...'}
+            if (errJson) {
+              if (errJson.errors && typeof errJson.errors === 'object') {
+                for (var f in errJson.errors) {
+                  // input ids use camelCase mapping (product_id -> inputProductId)
+                  var fid = 'input' + f.split('_').map(function (s, i) { return i === 0 ? s.charAt(0).toUpperCase() + s.slice(1) : s.charAt(0).toUpperCase() + s.slice(1); }).join('');
+                  // ex: 'product_id' => 'inputProductId'
+                  if ($id(fid)) {
+                    showCreateFieldError(fid, errJson.errors[f]);
+                  } else {
+                    var ce = $id('createError');
+                    if (ce) { ce.classList.remove('d-none'); ce.textContent += (errJson.errors[f] + '\n'); }
+                  }
+                }
+              } else if (errJson.message) {
+                errText = errJson.message;
+              } else {
+                errText = JSON.stringify(errJson);
+              }
+            }
+          } catch (e) {
+            try { errText = await res.text(); } catch (e2) { errText = 'HTTP ' + res.status; }
+          }
+          if (errText) {
+            var ce2 = $id('createError');
+            if (ce2) { ce2.classList.remove('d-none'); ce2.textContent = errText; }
+          }
+          throw new Error(errText || ('HTTP ' + res.status));
         }
+
+        // success
+        var created = null;
+        try { created = await res.json(); } catch (e) { created = null; }
+        if (createModal) {
+          try { createModal.hide(); } catch (e) { /* ignore */ }
+        } else {
+          // fallback: try to click close
+          var btnClose = document.querySelector('#createModal .btn-close');
+          if (btnClose) btnClose.click();
+        }
+
+        // refresh table
+        try {
+          if (typeof loadPromotions === 'function') loadPromotions();
+        } catch (e) { console.warn('reload failed', e); }
+
+        // show toast
+        showSuccessToast('Promotion created');
+
+      } catch (err) {
+        console.error('Create failed:', err);
+        // err message already placed into createError by above logic where possible
+        var ce3 = $id('createError');
+        if (ce3 && ce3.classList.contains('d-none')) {
+          ce3.classList.remove('d-none');
+          ce3.textContent = err.message || String(err);
+        }
+      } finally {
+        if (submitBtn) {
+          submitBtn.disabled = false;
+          submitBtn.textContent = 'Create';
+        }
+        createForm.removeAttribute('aria-busy');
+      }
+    });
+  })();
+
+  /* -------------------------
+     Delete modal: logic
+  ------------------------- */
+  (function initDeleteModal() {
+    var deleteModalEl = $id('deleteModal');
+    var deleteModal = null;
+    if (deleteModalEl && window.bootstrap && window.bootstrap.Modal) {
+      try {
+        deleteModal = new bootstrap.Modal(deleteModalEl, {});
+      } catch (e) {
+        console.warn('Bootstrap Modal init failed', e);
+      }
     }
 
-    // ****************************************
-    // Create a Promotion
-    // ****************************************
+    var currentDeleteId = null;
 
-    $("#create_promotion-btn").click(function () {
-        const payload = {
-            name: $("#promotion_name").val(),
-            promotion_type: $("#promotion_promotion_type").val(),
-            value: parseInt($("#promotion_value").val()),
-            product_id: parseInt($("#promotion_product_id").val()),
-            start_date: $("#promotion_start_date").val(),
-            end_date: $("#promotion_end_date").val()
-        };
+    // Event delegation for delete buttons
+    document.addEventListener('click', function (e) {
+      if (e.target && e.target.closest('.delete-btn')) {
+        var btn = e.target.closest('.delete-btn');
+        var id = btn.getAttribute('data-id');
+        var name = btn.getAttribute('data-name');
 
-        if (!payload.name || !payload.promotion_type) {
-            flashMessage("Name and Promotion Type are required to create a promotion");
-            return;
+        // Set modal content
+        $id('deletePromotionId').textContent = id;
+        $id('deletePromotionName').textContent = name;
+        currentDeleteId = id;
+
+        // Show modal
+        if (deleteModal) {
+          deleteModal.show();
         }
-
-        flashMessage("");
-
-        $.ajax({
-            type: "POST",
-            url: "/promotions",
-            contentType: "application/json",
-            data: JSON.stringify(payload),
-        })
-            .done(function (res) {
-                update_form_data(res);
-                renderPromotionTable([res]);
-                flashMessage("Success");
-            })
-            .fail(function (res) {
-                handleFail(res, "Failed to create promotion");
-            });
+      }
     });
 
+    // Confirm delete button handler
+    var confirmBtn = $id('confirmDelete');
+    if (confirmBtn) {
+      confirmBtn.addEventListener('click', async function () {
+        if (!currentDeleteId) return;
 
-    // ****************************************
-    // Update a Promotion
-    // ****************************************
+        // Disable button
+        confirmBtn.disabled = true;
+        confirmBtn.textContent = 'Deleting...';
 
-    // $("#update_promotion-btn").click(function () {
+        try {
+          var res = await fetch('/promotions/' + currentDeleteId, {
+            method: 'DELETE',
+            credentials: 'same-origin'
+          });
 
-    //     let promotion_id = $("#promotion_id").val();
-    //     let name = $("#promotion_name").val();
-    //     let type = $("#promotion_promotion_type").val();
-    //     let value = $("#promotion_value").val();
-    //     let product_id = $("#promotion_product_id").val();
-    //     let start_date = $("#promotion_start_date").val();
-    //     let end_date = $("#promotion_end_date").val();
+          if (!res.ok) {
+            throw new Error('DELETE failed: ' + res.status);
+          }
 
-    //     let data = {
-    //         "name": name,
-    //         "promotion_type": type,
-    //         "value": parseInt(value),
-    //         "product_id": parseInt(product_id),
-    //         "start_date": start_date,
-    //         "end_date": end_date
-    //     };
+          // Close modal
+          if (deleteModal) {
+            deleteModal.hide();
+          }
 
-    //     $("#flash_message").empty();
+          // Show success toast
+          showSuccessToast('Promotion deleted');
 
-    //     let ajax = $.ajax({
-    //             type: "PUT",
-    //             url: `/promotions/${promotion_id}`,
-    //             contentType: "application/json",
-    //             data: JSON.stringify(data)
-    //         })
+          // Refresh table
+          if (typeof loadPromotions === 'function') loadPromotions();
 
-    //     ajax.done(function(res){
-    //         update_form_data(res)
-    //         flash_message("Success")
-    //     });
+        } catch (err) {
+          console.error('Delete failed:', err);
+          alert('Failed to delete promotion: ' + err.message);
+        } finally {
+          confirmBtn.disabled = false;
+          confirmBtn.textContent = 'Delete';
+          currentDeleteId = null;
+        }
+      });
+    }
+  })();
 
-    //     ajax.fail(function(res){
-    //         flash_message(res.responseJSON.message)
-    //     });
+  /* -------------------------
+     Edit modal: logic
+  ------------------------- */
+  (function initEditModal() {
+    var editModalEl = $id('editModal');
+    var editModal = null;
+    if (editModalEl && window.bootstrap && window.bootstrap.Modal) {
+      try {
+        editModal = new bootstrap.Modal(editModalEl, {});
+      } catch (e) {
+        console.warn('Bootstrap Modal init failed', e);
+      }
+    }
 
-    // });
+    var currentEditId = null;
 
-    // ****************************************
-    // Retrieve a Promotion
-    // ****************************************
+    // Clear edit form errors
+    function clearEditFieldErrors() {
+      var ids = ['editName', 'editType', 'editValue', 'editProductId', 'editStart', 'editEnd'];
+      ids.forEach(function (id) {
+        var el = $id(id);
+        if (el) {
+          el.classList.remove('is-invalid');
+        }
+      });
+      var ee = $id('editError');
+      if (ee) { ee.classList.add('d-none'); ee.textContent = ''; }
+    }
 
-    // $("#retrieve_promotion-btn").click(function () {
+    function showEditFieldError(fieldId, message) {
+      var el = $id(fieldId);
+      if (!el) return;
+      el.classList.add('is-invalid');
+      var fb = el.nextElementSibling;
+      if (fb && fb.classList && fb.classList.contains('invalid-feedback')) {
+        fb.textContent = message;
+      } else {
+        var ee = $id('editError');
+        if (ee) { ee.classList.remove('d-none'); ee.textContent = message; }
+      }
+    }
 
-    //     let promotion_id = $("#promotion_id").val();
+    // Event delegation for edit buttons
+    document.addEventListener('click', function (e) {
+      if (e.target && e.target.closest('.edit-btn')) {
+        var btn = e.target.closest('.edit-btn');
+        var id = btn.getAttribute('data-id');
+        var promotionJson = btn.getAttribute('data-promotion');
 
-    //     $("#flash_message").empty();
+        try {
+          var promotion = JSON.parse(promotionJson);
 
-    //     let ajax = $.ajax({
-    //         type: "GET",
-    //         url: `/promotions/${promotion_id}`,
-    //         contentType: "application/json",
-    //         data: ''
-    //     })
+          // Populate form fields
+          $id('editId').value = id;
+          $id('editName').value = promotion.name || '';
+          $id('editType').value = promotion.promotion_type || '';
+          $id('editValue').value = promotion.value || '';
+          $id('editProductId').value = promotion.product_id || '';
+          $id('editStart').value = formatDateShort(promotion.start_date) || '';
+          $id('editEnd').value = formatDateShort(promotion.end_date) || '';
 
-    //     ajax.done(function(res){
-    //         update_form_data(res)
-    //         flash_message("Success")
-    //     });
+          currentEditId = id;
+          clearEditFieldErrors();
 
-    //     ajax.fail(function(res){
-    //         clear_form_data()
-    //         flash_message(res.responseJSON.message)
-    //     });
-
-    // });
-
-    // ****************************************
-    // Delete a Promotion
-    // ****************************************
-
-    $("#delete-btn").click(function () {
-
-        let promotion_id = $("#promotion_id").val();
-
-        $("#flash_message").empty();
-
-        let ajax = $.ajax({
-            type: "DELETE",
-            url: `/promotions/${promotion_id}`,
-            contentType: "application/json",
-            data: '',
-        })
-
-        ajax.done(function(res){
-            clear_form_data()
-            flashMessage("Promotion has been Deleted!")
-        });
-
-        ajax.fail(function(res){
-            handleFail(res, "Server error!")
-        });
+          // Show modal
+          if (editModal) {
+            editModal.show();
+          }
+        } catch (err) {
+          console.error('Failed to parse promotion data:', err);
+          alert('Failed to load promotion data');
+        }
+      }
     });
 
-    // ****************************************
-    // Clear the form
-    // ****************************************
+    // Handle modal close event
+    if (editModalEl) {
+      editModalEl.addEventListener('hidden.bs.modal', function () {
+        var form = $id('editForm');
+        if (form) {
+          form.reset();
+          form.classList.remove('was-validated');
+        }
+        clearEditFieldErrors();
+        currentEditId = null;
+      });
+    }
 
-    // $("#clear_form_fields-btn").click(function () {
-    //     $("#promotion_id").val("");
-    //     $("#flash_message").empty();
-    //     clear_form_data()
-    // });
+    // Handle form submission
+    var editForm = $id('editForm');
+    if (!editForm) return;
 
-    // ****************************************
-    // Search for a Promotion
-    // ****************************************
+    editForm.addEventListener('submit', async function (ev) {
+      ev.preventDefault();
+      ev.stopPropagation();
 
-    // $("#search_promotions-btn").click(function () {
+      editForm.classList.add('was-validated');
 
-    //     let name = $("#promotion_search_name").val();
-    //     let product_id = $("#promotion_search_product").val();
-    //     let promotion_type = $("#promotion_search_type").val();
+      // Basic HTML5 validity
+      if (!editForm.checkValidity()) {
+        return;
+      }
 
-    //     let queryString = ""
+      if (!currentEditId) {
+        alert('No promotion ID found');
+        return;
+      }
 
-    //     if (name) {
-    //         queryString += 'name=' + name
-    //     }
-    //     if (product_id) {
-    //         if (queryString.length > 0) {
-    //             queryString += '&product_id=' + product_id
-    //         } else {
-    //             queryString += 'product_id=' + product_id
-    //         }
-    //     }
-    //     if (promotion_type) {
-    //         if (queryString.length > 0) {
-    //             queryString += '&promotion_type=' + promotion_type
-    //         } else {
-    //             queryString += 'promotion_type=' + promotion_type
-    //         }
-    //     }
+      // Gather and coerce fields
+      clearEditFieldErrors();
 
-    //     $("#flash_message").empty();
+      var rawName = ($id('editName').value || '').trim();
+      var rawType = ($id('editType').value || '').trim();
+      var rawValue = $id('editValue').value;
+      var rawProductId = ($id('editProductId').value || '').trim();
+      var rawStart = $id('editStart').value || '';
+      var rawEnd = $id('editEnd').value || '';
 
-    //     let ajax = $.ajax({
-    //         type: "GET",
-    //         url: `/promotions?${queryString}`,
-    //         contentType: "application/json",
-    //         data: ''
-    //     })
+      // product_id: if provided, must be integer
+      var productId = null;
+      if (rawProductId !== '') {
+        var parsedPid = parseInt(rawProductId, 10);
+        if (Number.isNaN(parsedPid) || !Number.isFinite(parsedPid)) {
+          showEditFieldError('editProductId', 'Product ID must be an integer');
+          return;
+        }
+        productId = parsedPid;
+      } else {
+        productId = null;
+      }
 
-    //     ajax.done(function(res){
-    //         $("#search_results").empty();
-    //         let table = '<table class="table table-striped" cellpadding="10">'
-    //         table += '<thead><tr>'
-    //         table += '<th class="col-md-2">ID</th>'
-    //         table += '<th class="col-md-2">Name</th>'
-    //         table += '<th class="col-md-2">Type</th>'
-    //         table += '<th class="col-md-2">Value</th>'
-    //         table += '<th class="col-md-2">Product ID</th>'
-    //         table += '<th class="col-md-2">Start Date</th>'
-    //         table += '<th class="col-md-2">End Date</th>'
-    //         table += '</tr></thead><tbody>'
-    //         let firstPromotion = "";
-    //         for(let i = 0; i < res.length; i++) {
-    //             let promotion = res[i];
-    //             table +=  `<tr id="row_${i}"><td>${promotion.id}</td><td>${promotion.name}</td><td>${promotion.promotion_type}</td><td>${promotion.value}</td><td>${promotion.product_id}</td><td>${promotion.start_date}</td><td>${promotion.end_date}</td></tr>`;
-    //             if (i == 0) {
-    //                 firstPromotion = promotion;
-    //             }
-    //         }
-    //         table += '</tbody></table>';
-    //         $("#search_results").append(table);
+      // value coercion
+      var valueNum = Number(rawValue);
+      if (!Number.isFinite(valueNum) || valueNum < 0) {
+        showEditFieldError('editValue', 'Value must be a non-negative number');
+        return;
+      }
 
-    //         // copy the first result to the form
-    //         if (firstPromotion != "") {
-    //             update_form_data(firstPromotion)
-    //         }
+      // normalize dates
+      var startDate = normalizeDateInput(rawStart);
+      var endDate = normalizeDateInput(rawEnd);
 
-    //         flash_message("Success")
-    //     });
+      var payload = {
+        name: rawName || null,
+        promotion_type: rawType || null,
+        value: valueNum,
+        product_id: productId,
+        start_date: startDate,
+        end_date: endDate
+      };
 
-    //     ajax.fail(function(res){
-    //         flash_message(res.responseJSON.message)
-    //     });
+      // UI: disable submit
+      var submitBtn = $id('editSubmit');
+      if (submitBtn) {
+        submitBtn.disabled = true;
+        submitBtn.textContent = 'Updating...';
+      }
+      editForm.setAttribute('aria-busy', 'true');
 
-    // });
+      try {
+        var res = await fetch('/promotions/' + currentEditId, {
+          method: 'PUT',
+          headers: { 'Content-Type': 'application/json' },
+          credentials: 'same-origin',
+          body: JSON.stringify(payload)
+        });
 
-    // ****************************************
-    // List All Promotions
-    // ****************************************
+        if (!res.ok) {
+          // Try to parse structured errors
+          var errText = '';
+          try {
+            var errJson = await res.json();
+            if (errJson) {
+              if (errJson.errors && typeof errJson.errors === 'object') {
+                for (var f in errJson.errors) {
+                  var fid = 'edit' + f.split('_').map(function (s) { return s.charAt(0).toUpperCase() + s.slice(1); }).join('');
+                  if ($id(fid)) {
+                    showEditFieldError(fid, errJson.errors[f]);
+                  } else {
+                    var ee = $id('editError');
+                    if (ee) { ee.classList.remove('d-none'); ee.textContent += (errJson.errors[f] + '\n'); }
+                  }
+                }
+              } else if (errJson.message) {
+                errText = errJson.message;
+              } else {
+                errText = JSON.stringify(errJson);
+              }
+            }
+          } catch (e) {
+            try { errText = await res.text(); } catch (e2) { errText = 'HTTP ' + res.status; }
+          }
+          if (errText) {
+            var ee2 = $id('editError');
+            if (ee2) { ee2.classList.remove('d-none'); ee2.textContent = errText; }
+          }
+          throw new Error(errText || ('HTTP ' + res.status));
+        }
 
-    // $("#list_promotions-btn").click(function () {
+        // Success
+        var updated = null;
+        try { updated = await res.json(); } catch (e) { updated = null; }
+        if (editModal) {
+          try { editModal.hide(); } catch (e) { /* ignore */ }
+        } else {
+          var btnClose = document.querySelector('#editModal .btn-close');
+          if (btnClose) btnClose.click();
+        }
 
-    //     $("#flash_message").empty();
+        // Refresh table
+        try {
+          if (typeof loadPromotions === 'function') loadPromotions();
+        } catch (e) { console.warn('reload failed', e); }
 
-    //     let ajax = $.ajax({
-    //         type: "GET",
-    //         url: "/promotions",
-    //         contentType: "application/json",
-    //         data: ''
-    //     })
+        // Show toast
+        showSuccessToast('Promotion updated');
 
-    //     ajax.done(function(res){
-    //         $("#search_results").empty();
-    //         let table = '<table class="table table-striped" cellpadding="10">'
-    //         table += '<thead><tr>'
-    //         table += '<th class="col-md-2">ID</th>'
-    //         table += '<th class="col-md-2">Name</th>'
-    //         table += '<th class="col-md-2">Type</th>'
-    //         table += '<th class="col-md-2">Value</th>'
-    //         table += '<th class="col-md-2">Product ID</th>'
-    //         table += '<th class="col-md-2">Start Date</th>'
-    //         table += '<th class="col-md-2">End Date</th>'
-    //         table += '</tr></thead><tbody>'
-    //         let firstPromotion = "";
-    //         for(let i = 0; i < res.length; i++) {
-    //             let promotion = res[i];
-    //             table +=  `<tr id="row_${i}"><td>${promotion.id}</td><td>${promotion.name}</td><td>${promotion.promotion_type}</td><td>${promotion.value}</td><td>${promotion.product_id}</td><td>${promotion.start_date}</td><td>${promotion.end_date}</td></tr>`;
-    //             if (i == 0) {
-    //                 firstPromotion = promotion;
-    //             }
-    //         }
-    //         table += '</tbody></table>';
-    //         $("#search_results").append(table);
+      } catch (err) {
+        console.error('Update failed:', err);
+        var ee3 = $id('editError');
+        if (ee3 && ee3.classList.contains('d-none')) {
+          ee3.classList.remove('d-none');
+          ee3.textContent = err.message || String(err);
+        }
+      } finally {
+        if (submitBtn) {
+          submitBtn.disabled = false;
+          submitBtn.textContent = 'Update';
+        }
+        editForm.removeAttribute('aria-busy');
+      }
+    });
+  })();
 
-    //         // copy the first result to the form
-    //         if (firstPromotion != "") {
-    //             update_form_data(firstPromotion)
-    //         }
+  /* -------------------------
+     Filters: Event Handlers
+  ------------------------- */
+  (function initFilters() {
+    var searchInput = $id('searchInput');
+    var filterPills = document.querySelectorAll('.filter-pill');
+    var filterType = $id('filterType');
+    var filterProductId = $id('filterProductId');
+    var btnClearFilters = $id('btnClearFilters');
 
-    //         flash_message("Success")
-    //     });
+    var searchTimeout = null;
 
-    //     ajax.fail(function(res){
-    //         flash_message(res.responseJSON.message)
-    //     });
+    // Initialize filters from URL on page load
+    function initFromUrl() {
+      var urlParams = new URLSearchParams(window.location.search);
 
-    // });
+      if (urlParams.has('active')) {
+        var activeVal = urlParams.get('active');
+        if (activeVal === 'true') {
+          currentFilters.active = 'active';
+          filterPills.forEach(function (pill) {
+            pill.classList.remove('active');
+            if (pill.getAttribute('data-filter') === 'active') {
+              pill.classList.add('active');
+            }
+          });
+        } else if (activeVal === 'false') {
+          currentFilters.active = 'inactive';
+          filterPills.forEach(function (pill) {
+            pill.classList.remove('active');
+            if (pill.getAttribute('data-filter') === 'inactive') {
+              pill.classList.add('active');
+            }
+          });
+        }
+      } else if (urlParams.has('name')) {
+        currentFilters.name = urlParams.get('name');
+        if (searchInput) searchInput.value = currentFilters.name;
+      } else if (urlParams.has('promotion_type')) {
+        currentFilters.type = urlParams.get('promotion_type');
+        if (filterType) filterType.value = currentFilters.type;
+      } else if (urlParams.has('product_id')) {
+        currentFilters.productId = urlParams.get('product_id');
+        if (filterProductId) filterProductId.value = currentFilters.productId;
+      }
+    }
 
-    // ****************************************
-    // Deactivate a Promotion (ACTION)
-    // ****************************************
+    // Initialize from URL
+    initFromUrl();
 
-    // $("#deactivate_promotion-btn").click(function () {
+    // Handle browser back/forward buttons
+    window.addEventListener('popstate', function () {
+      // Reset filters
+      currentFilters.active = null;
+      currentFilters.name = null;
+      currentFilters.type = null;
+      currentFilters.productId = null;
 
-    //     let promotion_id = $("#promotion_action_id").val();
+      // Reinitialize from URL
+      initFromUrl();
 
-    //     $("#flash_message").empty();
+      // Apply filters without updating URL (already changed by popstate)
+      var queryParams = window.location.search.substring(1);
+      loadPromotions(queryParams);
+    });
 
-    //     let ajax = $.ajax({
-    //         type: "PUT",
-    //         url: `/promotions/${promotion_id}/deactivate`,
-    //         contentType: "application/json",
-    //         data: ''
-    //     })
+    function applyFilters() {
+      var queryParams = '';
 
-    //     ajax.done(function(res){
-    //         update_form_data(res)
-    //         flash_message("Promotion has been deactivated!")
-    //     });
+      // Priority: active > name > type > product_id (backend priority)
+      if (currentFilters.active && currentFilters.active !== 'all') {
+        // active: true or false
+        if (currentFilters.active === 'active') {
+          queryParams = 'active=true';
+        } else if (currentFilters.active === 'inactive') {
+          queryParams = 'active=false';
+        }
+      } else if (currentFilters.name && currentFilters.name.trim()) {
+        queryParams = 'name=' + encodeURIComponent(currentFilters.name.trim());
+      } else if (currentFilters.type) {
+        queryParams = 'promotion_type=' + encodeURIComponent(currentFilters.type);
+      } else if (currentFilters.productId) {
+        queryParams = 'product_id=' + encodeURIComponent(currentFilters.productId);
+      }
 
-    //     ajax.fail(function(res){
-    //         flash_message(res.responseJSON.message)
-    //     });
+      // Update browser URL
+      var newUrl = window.location.pathname;
+      if (queryParams) {
+        newUrl += '?' + queryParams;
+      }
+      window.history.pushState({}, '', newUrl);
 
-    // });
+      loadPromotions(queryParams);
+    }
 
-})
+    // Search input (debounced)
+    if (searchInput) {
+      searchInput.addEventListener('input', function (e) {
+        clearTimeout(searchTimeout);
+        searchTimeout = setTimeout(function () {
+          currentFilters.name = e.target.value;
+          // Clear other filters when searching by name
+          if (currentFilters.name && currentFilters.name.trim()) {
+            currentFilters.active = null;
+            currentFilters.type = null;
+            currentFilters.productId = null;
+            // Reset UI
+            filterPills.forEach(function (pill) { pill.classList.remove('active'); });
+            if (filterType) filterType.value = '';
+            if (filterProductId) filterProductId.value = '';
+          }
+          applyFilters();
+        }, 300);
+      });
+    }
+
+    // Status pills
+    filterPills.forEach(function (pill) {
+      pill.addEventListener('click', function () {
+        var filter = this.getAttribute('data-filter');
+
+        // Update active state
+        filterPills.forEach(function (p) { p.classList.remove('active'); });
+        this.classList.add('active');
+
+        currentFilters.active = filter;
+
+        // Clear other filters
+        if (filter !== 'all') {
+          currentFilters.name = null;
+          currentFilters.type = null;
+          currentFilters.productId = null;
+          // Reset UI
+          if (searchInput) searchInput.value = '';
+          if (filterType) filterType.value = '';
+          if (filterProductId) filterProductId.value = '';
+        }
+
+        applyFilters();
+      });
+    });
+
+    // Type dropdown
+    if (filterType) {
+      filterType.addEventListener('change', function (e) {
+        currentFilters.type = e.target.value;
+
+        // Clear other filters when selecting type
+        if (currentFilters.type) {
+          currentFilters.active = null;
+          currentFilters.name = null;
+          currentFilters.productId = null;
+          // Reset UI
+          filterPills.forEach(function (pill) { pill.classList.remove('active'); });
+          if (searchInput) searchInput.value = '';
+          if (filterProductId) filterProductId.value = '';
+        }
+
+        applyFilters();
+      });
+    }
+
+    // Product ID input (debounced)
+    if (filterProductId) {
+      filterProductId.addEventListener('input', function (e) {
+        clearTimeout(searchTimeout);
+        searchTimeout = setTimeout(function () {
+          currentFilters.productId = e.target.value;
+
+          // Clear other filters when entering product ID
+          if (currentFilters.productId) {
+            currentFilters.active = null;
+            currentFilters.name = null;
+            currentFilters.type = null;
+            // Reset UI
+            filterPills.forEach(function (pill) { pill.classList.remove('active'); });
+            if (searchInput) searchInput.value = '';
+            if (filterType) filterType.value = '';
+          }
+
+          applyFilters();
+        }, 300);
+      });
+    }
+
+    // Clear filters button
+    if (btnClearFilters) {
+      btnClearFilters.addEventListener('click', function () {
+        // Reset all filters
+        currentFilters.active = null;
+        currentFilters.name = null;
+        currentFilters.type = null;
+        currentFilters.productId = null;
+
+        // Reset UI
+        if (searchInput) searchInput.value = '';
+        if (filterType) filterType.value = '';
+        if (filterProductId) filterProductId.value = '';
+        filterPills.forEach(function (pill) {
+          pill.classList.remove('active');
+          if (pill.getAttribute('data-filter') === 'all') {
+            pill.classList.add('active');
+          }
+        });
+
+        // Clear URL parameters
+        window.history.pushState({}, '', window.location.pathname);
+
+        // Reload all promotions
+        loadPromotions();
+      });
+    }
+  })();
+
+  // end of IIFE
+})();
